Now Graph has functions with class SFunction. SFunction is removed. Array2D is removed. GetID() and Vertex numeration problem. Search is completed.

+++ Find the shortest route/Path.cpp
@@ -10,7 +10,7 @@ Path::~Path()
 
 }
 
-size_t Path::getCost() const
+uint8_t Path::getCost() const
 {
 	return cost;
 }
